% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_scaling.R
\name{apply_nmar_scaling}
\alias{apply_nmar_scaling}
\title{Apply scaling to a matrix using a recipe}
\usage{
apply_nmar_scaling(matrix_to_scale, recipe)
}
\arguments{
\item{matrix_to_scale}{a numeric matrix with column names present in `recipe`.}

\item{recipe}{an object of class `nmar_scaling_recipe`.}
}
\value{
a matrix with each named column centered and scaled using the recipe
}
\description{
Apply scaling to a matrix using a recipe
}
\keyword{internal}
