% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_nmar_result_methods.R
\name{autoplot.nmar_result}
\alias{autoplot.nmar_result}
\title{Default ggplot2 autoplot for NMAR results}
\usage{
\method{autoplot}{nmar_result}(object, type = c("weights", "fitted", "constraints"), ...)
}
\arguments{
\item{object}{An object of class `nmar_result` or a subclass.}

\item{type}{One of "weights", "fitted", or "constraints".}

\item{...}{Ignored.}
}
\value{
A `ggplot` object.
}
\description{
Quick ggplot2 visualizations for result objects.
}
\keyword{internal}
