% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_nmar_result_methods.R
\name{autoplot}
\alias{autoplot}
\title{ggplot2 autoplot generic}
\usage{
autoplot(object, ...)
}
\arguments{
\item{object}{An object.}

\item{...}{Passed to methods.}
}
\description{
Generic for autoplot; methods provide plotting for NMAR results.
}
\keyword{internal}
