% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_bootstrap.R
\name{bootstrap_variance.svyrep.design}
\alias{bootstrap_variance.svyrep.design}
\title{Replicate-weight designs not supported}
\usage{
\method{bootstrap_variance}{svyrep.design}(data, estimator_func, point_estimate, ...)
}
\description{
Replicate-weight designs not supported
}
\keyword{internal}
