% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_nmar_result_methods.R
\name{coef.nmar_result}
\alias{coef.nmar_result}
\title{Default coefficients for NMAR results}
\usage{
\method{coef}{nmar_result}(object, ...)
}
\arguments{
\item{object}{An `nmar_result` object.}

\item{...}{Ignored.}
}
\value{
A named numeric vector or `NULL`.
}
\description{
Returns missingness-model coefficients.
}
\keyword{result_param}
