% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_scaling.R
\name{compute_weighted_stats}
\alias{compute_weighted_stats}
\title{Compute mean and standard deviation}
\usage{
compute_weighted_stats(values, weights = NULL)
}
\description{
Compute mean and standard deviation
}
\keyword{internal}
