% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_nmar_result_methods.R
\name{confint.nmar_result}
\alias{confint.nmar_result}
\title{Wald confidence interval for base NMAR results}
\usage{
\method{confint}{nmar_result}(object, parm, level = 0.95, ...)
}
\arguments{
\item{object}{An object of class `nmar_result`.}

\item{parm}{Ignored.}

\item{level}{Confidence level.}

\item{...}{Ignored.}
}
\description{
Wald confidence interval for base NMAR results
}
\keyword{result_param}
