% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engines_el_impl_dataframe.R
\name{el.data.frame}
\alias{el.data.frame}
\title{Empirical likelihood for data frames}
\usage{
\method{el}{data.frame}(
  data,
  formula,
  auxiliary_means = NULL,
  standardize = TRUE,
  trim_cap = Inf,
  control = list(),
  on_failure = c("return", "error"),
  variance_method = c("bootstrap", "none"),
  bootstrap_reps = 500,
  n_total = NULL,
  start = NULL,
  trace_level = 0,
  family = logit_family(),
  ...
)
}
\description{
Empirical likelihood for data frames
}
\keyword{internal}
