% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engines_el_impl_el.R
\name{el}
\alias{el}
\title{Empirical likelihood estimator}
\usage{
el(data, ...)
}
\arguments{
\item{data}{A \code{data.frame} or a \code{survey.design}.}

\item{...}{Passed to class-specific methods.}
}
\description{
Empirical likelihood estimator
}
\keyword{internal}
