% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engines_el_impl_survey.R
\name{el.survey.design}
\alias{el.survey.design}
\title{Empirical likelihood estimator for survey designs}
\usage{
\method{el}{survey.design}(
  data,
  formula,
  auxiliary_means = NULL,
  standardize = TRUE,
  strata_augmentation = TRUE,
  trim_cap = Inf,
  control = list(),
  on_failure = c("return", "error"),
  variance_method = c("bootstrap", "none"),
  bootstrap_reps = 500,
  n_total = NULL,
  start = NULL,
  trace_level = 0,
  family = logit_family(),
  ...
)
}
\description{
Empirical likelihood estimator for survey designs
}
\keyword{internal}
