% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engines_el_impl_utils_aux.R
\name{el_assert_no_offset}
\alias{el_assert_no_offset}
\title{Assert that terms object lacks offsets}
\usage{
el_assert_no_offset(terms_obj, label)
}
\description{
Assert that terms object lacks offsets
}
\keyword{internal}
