% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engines_el_impl_result.R
\name{el_build_result}
\alias{el_build_result}
\title{Build EL result object}
\usage{
el_build_result(
  core_results,
  inputs,
  call,
  formula,
  engine_name = "empirical_likelihood"
)
}
\description{
Build EL result object
}
\keyword{internal}
