% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engines_el_impl_core_helpers.R
\name{el_compute_diagnostics}
\alias{el_compute_diagnostics}
\title{Compute diagnostics}
\usage{
el_compute_diagnostics(
  estimates,
  equation_system_func,
  analytical_jac_func,
  post,
  respondent_weights,
  auxiliary_matrix_scaled,
  K_beta,
  K_aux,
  X_centered
)
}
\description{
Compute diagnostics
}
\keyword{internal}
