% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engines_el_impl_core_helpers.R
\name{el_compute_variance}
\alias{el_compute_variance}
\title{Variance driver}
\usage{
el_compute_variance(
  y_hat,
  full_data,
  formula,
  N_pop,
  variance_method,
  bootstrap_reps,
  standardize,
  trim_cap,
  on_failure,
  auxiliary_means,
  control,
  start,
  family
)
}
\description{
Variance driver
}
\keyword{internal}
