% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engines_el_impl_utils_math.R
\name{el_denominator}
\alias{el_denominator}
\title{Compute denominator}
\usage{
el_denominator(lambda_W, W, Xc_lambda, p_i, floor)
}
\arguments{
\item{lambda_W}{numeric scalar}

\item{W}{numeric scalar in (0,1)}

\item{Xc_lambda}{numeric vector (X_centered \%*\% lambda_x) or 0}

\item{p_i}{numeric vector of response probabilities}

\item{floor}{numeric scalar > 0, denominator floor}
}
\value{
list with denom, active, inv, inv_sq
}
\description{
Compute denominator
}
\keyword{internal}
