% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engines_el_impl_utils_math.R
\name{el_lambda_W}
\alias{el_lambda_W}
\title{Compute lambda_W from C_const and W}
\usage{
el_lambda_W(C_const, W)
}
\arguments{
\item{C_const}{numeric scalar: (N_pop / sum(d_resp) - 1)}

\item{W}{numeric scalar in (0,1)}
}
\description{
Compute lambda_W from C_const and W
}
\keyword{internal}
