% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engines_el_impl_logging.R
\name{el_log_detailed_diagnostics}
\alias{el_log_detailed_diagnostics}
\title{Log detailed diagnostics}
\usage{
el_log_detailed_diagnostics(
  verboser,
  beta_hat_unscaled,
  W_hat,
  lambda_W_hat,
  lambda_hat,
  denominator_hat
)
}
\description{
Log detailed diagnostics
}
\keyword{internal}
