% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engines_el_impl_logging.R
\name{el_log_solver_config}
\alias{el_log_solver_config}
\title{Log solver configuration}
\usage{
el_log_solver_config(verboser, control_top, final_control)
}
\description{
Log solver configuration
}
\keyword{internal}
