% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engines_el_impl_logging.R
\name{el_log_solver_result}
\alias{el_log_solver_result}
\title{Log solver termination status}
\usage{
el_log_solver_result(verboser, converged_success, solution, elapsed)
}
\description{
Log solver termination status
}
\keyword{internal}
