% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engines_el_impl_logging.R
\name{el_log_start_values}
\alias{el_log_start_values}
\title{Log starting values}
\usage{
el_log_start_values(verboser, init_beta, init_z, init_lambda)
}
\description{
Log starting values
}
\keyword{internal}
