% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engines_el_impl_logging.R
\name{el_log_trace}
\alias{el_log_trace}
\title{Log a short trace message with the chosen level}
\usage{
el_log_trace(verboser, trace_level)
}
\description{
Log a short trace message with the chosen level
}
\keyword{internal}
