% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engines_el_impl_logging.R
\name{el_log_variance_header}
\alias{el_log_variance_header}
\title{Log variance header and result}
\usage{
el_log_variance_header(verboser, variance_method, bootstrap_reps)
}
\description{
Log variance header and result
}
\keyword{internal}
