% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engines_el_impl_logging.R
\name{el_log_weight_diagnostics}
\alias{el_log_weight_diagnostics}
\title{Log weight diagnostics}
\usage{
el_log_weight_diagnostics(verboser, W_hat, weights, trimmed_fraction)
}
\description{
Log weight diagnostics
}
\keyword{internal}
