% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engines_el_impl_survey.R
\name{el_rescale_survey_design_weights}
\alias{el_rescale_survey_design_weights}
\title{Rescale survey design weights in place}
\usage{
el_rescale_survey_design_weights(design, scale_factor)
}
\description{
Rescale survey design weights in place
}
\keyword{internal}
