% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engines_el_impl_utils_aux.R
\name{el_validate_matrix}
\alias{el_validate_matrix}
\title{Validate matrix columns for NA and zero variance}
\usage{
el_validate_matrix(
  mat,
  allow_na,
  label,
  severity,
  row_map = NULL,
  scope_note = NULL,
  plural_label = FALSE
)
}
\description{
Validate matrix columns for NA and zero variance
}
\keyword{internal}
