% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_weights.R
\name{enforce_nonneg_weights}
\alias{enforce_nonneg_weights}
\title{Enforce (near-)nonnegativity of weights}
\usage{
enforce_nonneg_weights(weights, tol = 1e-08)
}
\description{
Enforce (near-)nonnegativity of weights
}
\keyword{internal}
