% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_nmar_engine_methods.R
\name{engine_config}
\alias{engine_config}
\title{Extract engine configuration}
\usage{
engine_config(x)
}
\arguments{
\item{x}{An object inheriting from class `nmar_engine`.}
}
\value{
A named list of configuration fields.
}
\description{
Extract engine configuration
}
\keyword{engine_view}
