% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engines_exptilt_engine.R
\name{exptilt}
\alias{exptilt}
\title{Exponential Tilting Engine for NMAR Estimation}
\usage{
exptilt(
  prob_model_type = get_json_param_info(all_schemas, "exptilt",
    "prob_model_type")$default,
  y_dens = get_json_param_info(all_schemas, "exptilt", "y_dens")$default,
  tol_value = get_json_param_info(all_schemas, "exptilt", "tol_value")$default,
  min_iter = get_json_param_info(all_schemas, "exptilt", "min_iter")$default,
  max_iter = get_json_param_info(all_schemas, "exptilt", "max_iter")$default,
  optim_method = get_json_param_info(all_schemas, "exptilt", "optim_method")$default
)
}
\arguments{
\item{prob_model_type}{Probability model for data missingness ("logit" or "probit")}

\item{y_dens}{Density model for outcome variable ("normal" or "gamma")}

\item{tol_value}{Tolerance in convergence (default = 0.00001)}

\item{min_iter}{Minimum number of iterations (default = 10)}

\item{max_iter}{Maximum number of iterations (default = 100)}

\item{optim_method}{Optimization method ("Newton" or "Broyden")}
}
\value{
Engine configuration object of class `nmar_engine_exptilt`
}
\description{
Exponential Tilting Engine for NMAR Estimation
}
