% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_numerics.R
\name{extract_nleqslv_top}
\alias{extract_nleqslv_top}
\title{Extract top-level nleqslv arguments from a control-like list}
\usage{
extract_nleqslv_top(ctrl)
}
\description{
Extract top-level nleqslv arguments from a control-like list
}
\keyword{internal}
