% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_nmar_result_methods.R
\name{fitted.nmar_result}
\alias{fitted.nmar_result}
\title{Default fitted values for NMAR results}
\usage{
\method{fitted}{nmar_result}(object, ...)
}
\arguments{
\item{object}{An `nmar_result` object.}

\item{...}{Ignored.}
}
\value{
A numeric vector (possibly length 0).
}
\description{
Returns fitted response probabilities if available.
}
\keyword{result_param}
