% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_nmar_result_methods.R
\name{formula.nmar_result}
\alias{formula.nmar_result}
\title{Default formula for NMAR results}
\usage{
\method{formula}{nmar_result}(x, ...)
}
\arguments{
\item{x}{An `nmar_result` object.}

\item{...}{Ignored.}
}
\value{
A formula or `NULL`.
}
\description{
Returns the estimation formula if available.
}
\keyword{result_param}
