% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_nmar_result_methods.R
\name{glance.nmar_result}
\alias{glance.nmar_result}
\title{Glance summary for NMAR results}
\usage{
\method{glance}{nmar_result}(x, ...)
}
\arguments{
\item{x}{An object of class `nmar_result`.}

\item{...}{Ignored.}
}
\description{
One-row diagnostics for NMAR fits.
}
\keyword{result_view}
