% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTAR.sim.R
\name{MTAR.sim}
\alias{MTAR.sim}
\title{MTAR- Simulation}
\usage{
MTAR.sim(
  m = 2,
  n = 3,
  p = 1,
  regimes = 2,
  nsim = 1000,
  constant = TRUE,
  A,
  B,
  M = NULL,
  threshold = 0.3,
  st_type = c("trend", "autoregressive"),
  Etype = c("identity", "normal"),
  sdEt = 0.2
)
}
\arguments{
\item{m}{number of rows of the matrix-variate time series}

\item{n}{number of columns of the matrix-variate time series}

\item{p}{number of lags}

\item{regimes}{number of regimes}

\item{nsim}{number of simulated data}

\item{constant}{\code{TRUE} or \code{FALSE} to include or not a constant matrix.}

\item{A}{\code{list} of values for the row-wise coefficient \code{(m×m)} matrices, one list for each lag containing a list of matrices for each regime.}

\item{B}{\code{list} of values for the column-wise coefficient \code{(n×n)} matrices, one list for each lag containing a list of matrices for each regime.}

\item{M}{\code{list} of values for the constant \code{(m×n)} matrices, a list of matrices for each regime.}

\item{threshold}{vector of thresholds used for the simulation equal to the number of regimes minus 1}

\item{st_type}{character string, specifying the type of the threshold used for the simulation. \describe{
 \item{\code{"trend",}}{\code{st = t/T} as a transition variable}
 \item{\code{"autoregressive",}}{an exogenous variable with an autoregressive dynamics}
}}

\item{Etype}{character string, specifying the type of the threshold used for the simulation. \describe{
 \item{\code{"identity",}}{two identity matrices are used as row-wise and column-wise covariance matrices}
 \item{\code{"normal",}}{the \code{mnxmn} covariance matrix is sampled from a normal distribution}
}}

\item{sdEt}{standard deviation for the sampling of the covariance matrix}

\item{burnin}{number of data to be burned}
}
\value{
return a list containing the following:\describe{
\item{\code{data}}{an \code{m×n×T} array with data simulated from a MTAR model}
\item{\code{st}}{the transition variable used to simulate the data}
\item{\code{threshold}}{a vector of thresholds equal to the number of regimes minus 1}
\item{\code{datavec}}{an \code{mn×T} matrix with matrix data stacked in a vector}
}
}
\description{
This function allows the user to simulate the data from a matrix-variate threshold autoregressive model with \emph{m} regimes.
}
\examples{
m = 3
n = 2
nsim = 1000
A = list(list(matrix(0.05, m, m), matrix(0.10, m, m), matrix(0.15, m, m)))
B = list(list(matrix(0.05, n, n), matrix(0.10, n, n), matrix(0.15, n, n)))
M = list(matrix(1, m, n), matrix(1.5, m, n), matrix(2, m, n))
simuldata = MTAR.sim(m=m, n=n, p = 1, nsim = 1000, 
regimes = 3, threshold = c(0.3, 0.7), st_type = 'trend', 
Etype = 'normal', A = A, B= B, M = M, sdEt = 0.1)

plot.ts(simuldata$data[1,1,])
}
\references{
Liu X. and Chen E.Y. (2022), Identification and estimation of threshold matrix-variate factor models. \emph{Scandinavian Journal of Statistics}. 49: 1383-1417
}
\author{
Andrea Bucci
}
\keyword{MTAR.sim}
