% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engines_el_impl_constructors.R
\name{new_nmar_result_el}
\alias{new_nmar_result_el}
\title{Construct EL Result Object}
\usage{
new_nmar_result_el(
  y_hat,
  se,
  weights,
  coefficients,
  vcov,
  converged,
  diagnostics,
  inputs,
  nmar_scaling_recipe,
  fitted_values,
  call,
  formula = NULL
)
}
\description{
Construct EL Result Object
}
\keyword{internal}
