% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_nmar_result_utils.R
\name{nmar_fmt_num}
\alias{nmar_fmt_num}
\title{Format a number with fixed decimal places using nmar.digits}
\usage{
nmar_fmt_num(x, digits = nmar_get_digits())
}
\description{
Format a number with fixed decimal places using nmar.digits
}
\keyword{internal}
