% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_nmar_result_utils.R
\name{nmar_get_digits}
\alias{nmar_get_digits}
\title{Resolve global digits setting for printing}
\usage{
nmar_get_digits()
}
\description{
Resolve global digits setting for printing
}
\keyword{internal}
