% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_nmar_result_utils.R
\name{nmar_result_get_estimate}
\alias{nmar_result_get_estimate}
\title{Internal helpers for nmar_result objects}
\usage{
nmar_result_get_estimate(x)
}
\description{
Internal helpers for nmar_result objects
}
\keyword{internal}
