% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_nmar_engine_methods.R
\name{print.nmar_engine}
\alias{print.nmar_engine}
\title{Print method for engines}
\usage{
\method{print}{nmar_engine}(x, ...)
}
\arguments{
\item{x}{An engine object inheriting from `nmar_engine`.}

\item{...}{Unused.}
}
\value{
`x`, invisibly.
}
\description{
Compact summary for `nmar_engine` objects.
}
\keyword{engine_view}
