% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engines_el_s3.R
\name{print.nmar_result_el}
\alias{print.nmar_result_el}
\title{Print method for EL results}
\usage{
\method{print}{nmar_result_el}(x, ...)
}
\arguments{
\item{x}{An object of class `nmar_result_el`.}

\item{...}{Ignored.}
}
\description{
Compact print for objects of class `nmar_result_el`.
}
\keyword{result_view}
