% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engines_induced_logit_s3.R
\name{print.nmar_result_induced_logit}
\alias{print.nmar_result_induced_logit}
\title{Print method for induced-logit results}
\usage{
\method{print}{nmar_result_induced_logit}(x, ...)
}
\arguments{
\item{x}{An object of class `nmar_result_induced_logit`.}

\item{...}{Ignored.}
}
\value{
`x`, invisibly.
}
\description{
Print for objects of class `nmar_result_induced_logit`.
}
\keyword{result_view}
