% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_families.R
\name{probit_family}
\alias{probit_family}
\title{Construct probit response family}
\usage{
probit_family()
}
\value{
A list with components \code{name}, \code{linkinv}, \code{mu.eta},
\code{d2mu.deta2}, and \code{score_eta}.
}
\description{
Construct probit response family
}
\keyword{internal}
