% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engines_el_run_engine.R
\name{run_engine.nmar_engine_el}
\alias{run_engine.nmar_engine_el}
\title{Run method for EL engine}
\usage{
\method{run_engine}{nmar_engine_el}(engine, formula, data, trace_level = 0)
}
\description{
Run method for EL engine
}
\keyword{internal}
