% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_nmar_result_methods.R
\name{se}
\alias{se}
\title{Extract standard error for NMAR results}
\usage{
se(object, ...)
}
\arguments{
\item{object}{An `nmar_result` or subclass.}

\item{...}{Ignored.}
}
\value{
Numeric scalar.
}
\description{
Returns the standard error of the primary mean estimate.
}
\keyword{result_param}
