% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_results.R
\name{summary.nmar_result}
\alias{summary.nmar_result}
\title{Summary method for nmar_result}
\usage{
\method{summary}{nmar_result}(object, ...)
}
\arguments{
\item{object}{nmar_result object}

\item{...}{Additional parameters}
}
\description{
Summary method for nmar_result
}
