% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engines_el_s3.R
\name{summary.nmar_result_el}
\alias{summary.nmar_result_el}
\title{Summary method for EL results}
\usage{
\method{summary}{nmar_result_el}(object, ...)
}
\arguments{
\item{object}{An object of class \code{nmar_result_el}.}

\item{...}{Ignored.}
}
\value{
An object of class \code{summary_nmar_result_el}.
}
\description{
Summarize estimation, standard error and missingness-model coefficients.
}
\keyword{result_view}
