% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engines_exptilt_s3.R
\name{summary.nmar_result_exptilt}
\alias{summary.nmar_result_exptilt}
\title{Summary method for Exponential Tilting results (engine-specific)}
\usage{
\method{summary}{nmar_result_exptilt}(object, conf.level = 0.95, ...)
}
\arguments{
\item{object}{An object of class `nmar_result_exptilt`.}

\item{conf.level}{Confidence level for confidence interval (default 0.95).}

\item{...}{Ignored.}
}
\value{
An object of class `summary_nmar_result_exptilt`.
}
\description{
Summarize estimation, standard error and model coefficients.
}
\keyword{result_view}
