% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engines_induced_logit_s3.R
\name{summary.nmar_result_induced_logit}
\alias{summary.nmar_result_induced_logit}
\title{Summary method for induced-logit results}
\usage{
\method{summary}{nmar_result_induced_logit}(object, ...)
}
\arguments{
\item{object}{An object of class `nmar_result_induced_logit`.}

\item{...}{Ignored.}
}
\value{
An object of class `summary_nmar_result_induced_logit`.
}
\description{
Summarize estimation, standard error and missingness-model coefficients.
}
\keyword{result_view}
