% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_weights.R
\name{trim_weights}
\alias{trim_weights}
\title{Trim weights by capping and proportional redistribution (water-filling)}
\usage{
trim_weights(weights, cap, tol = 1e-12, warn_tol = 1e-08)
}
\description{
Trim weights by capping and proportional redistribution (water-filling)
}
\keyword{internal}
