% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_numerics.R
\name{validate_nleqslv_top}
\alias{validate_nleqslv_top}
\title{Validate top-level nleqslv arguments and coerce invalid to defaults}
\usage{
validate_nleqslv_top(top)
}
\description{
Validate top-level nleqslv arguments and coerce invalid to defaults
}
\keyword{internal}
