% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engines_el_engine.R
\name{validate_nmar_engine_el}
\alias{validate_nmar_engine_el}
\title{Validate EL Engine Settings}
\usage{
validate_nmar_engine_el(engine)
}
\description{
Validate EL Engine Settings
}
\keyword{internal}
