% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_validate_arguments.R
\docType{data}
\name{validator}
\alias{validator}
\title{Custom Validation Module}
\format{
An object of class \code{environment} of length 13.
}
\usage{
validator
}
\description{
This module provides a set of internal helper functions for argument validation.
The functions are designed to throw a clear error message if a validation check fails.
}
\keyword{internal}
