% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_scaling.R
\name{apply_nmar_scaling}
\alias{apply_nmar_scaling}
\title{Apply scaling to a matrix using a recipe}
\usage{
apply_nmar_scaling(matrix_to_scale, recipe)
}
\arguments{
\item{matrix_to_scale}{A numeric matrix with column names present in
\code{recipe}.}

\item{recipe}{An object of class \code{nmar_scaling_recipe}.}
}
\value{
A numeric matrix with each column centered and scaled using
  \code{recipe}.
}
\description{
Apply scaling to a matrix using a recipe
}
\keyword{internal}
