% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_bootstrap.R
\name{bootstrap_variance.default}
\alias{bootstrap_variance.default}
\title{Default method dispatch (internal safety net)}
\usage{
bootstrap_variance.default(data, estimator_func, point_estimate, ...)
}
\description{
Default method dispatch (internal safety net)
}
\keyword{internal}
