% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engines_el_impl_el.R
\name{el}
\alias{el}
\title{Empirical likelihood estimator}
\usage{
el(data, ...)
}
\arguments{
\item{data}{A `data.frame` or a `survey.design`.}

\item{...}{Passed to class-specific methods.}
}
\description{
Generic for the empirical likelihood (EL) estimator under NMAR.
  Methods are provided for `data.frame` and `survey.design`.
}
\seealso{
`el.data.frame()`, `el.survey.design()`, `el_engine()`
}
\keyword{internal}
