% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engines_el_impl_el.R
\name{el}
\alias{el}
\title{Empirical likelihood estimator}
\usage{
el(data, ...)
}
\arguments{
\item{data}{A \code{data.frame} or a \code{survey.design}.}

\item{...}{Passed to class-specific methods.}
}
\description{
Generic for the empirical likelihood (EL) estimator under NMAR.
  Methods are provided for \code{data.frame} and \code{survey.design}.
}
\seealso{
\code{\link{el_engine}}
}
\keyword{internal}
