% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engines_el_impl_core_helpers.R
\name{el_build_start}
\alias{el_build_start}
\title{Build starting values for the EL solver (beta, z, lambda)}
\usage{
el_build_start(
  missingness_model_matrix_scaled,
  auxiliary_matrix_scaled,
  nmar_scaling_recipe,
  start,
  N_pop,
  respondent_weights
)
}
\description{
Build starting values for the EL solver (beta, z, lambda)
}
\keyword{internal}
