% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engines_el_impl_logging.R
\name{el_log_data_prep}
\alias{el_log_data_prep}
\title{Log data prep summary}
\usage{
el_log_data_prep(
  verboser,
  outcome_var,
  family_name,
  K_beta,
  K_aux,
  auxiliary_names,
  standardize,
  is_survey,
  N_pop,
  n_resp_weighted
)
}
\description{
Log data prep summary
}
\keyword{internal}
