% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_nmar_engine_methods.R
\name{engine_name}
\alias{engine_name}
\title{Canonical engine name}
\usage{
engine_name(x)
}
\arguments{
\item{x}{An object inheriting from class `nmar_engine`.}
}
\value{
A single character string, e.g. "empirical_likelihood".
}
\description{
Returns identifier for an engine object.
}
\keyword{engine_view}
