% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engines_exptilt_engine.R
\name{exptilt}
\alias{exptilt}
\title{Exponential Tilting Engine for NMAR Estimation}
\usage{
exptilt(
  prob_model_type = "logit",
  y_dens = "normal",
  tol_value = 1e-05,
  min_iter = 10,
  max_iter = 100,
  optim_method = "Newton"
)
}
\arguments{
\item{prob_model_type}{Probability model for data missingness ("logit" or "probit")}

\item{y_dens}{Density model for outcome variable ("normal" or "gamma")}

\item{tol_value}{Tolerance in convergence (default = 0.00001)}

\item{min_iter}{Minimum number of iterations (default = 10)}

\item{max_iter}{Maximum number of iterations (default = 100)}

\item{optim_method}{Optimization method ("Newton" or "Broyden")}
}
\value{
Engine configuration object of class `nmar_engine_exptilt`
}
\description{
Exponential Tilting Engine for NMAR Estimation
}
