% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/engines_exptilt-nonparam_impl_exptilt_nonparam.R
\name{exptilt_nonparam}
\alias{exptilt_nonparam}
\title{Nonparametric Exponential Tilting (Internal Generic)}
\usage{
exptilt_nonparam(data, ...)
}
\arguments{
\item{data}{A data.frame or survey.design object}

\item{...}{Other arguments passed to methods}
}
\value{
An engine-specific NMAR result object for the nonparametric
  exponential tilting estimator.
}
\description{
Nonparametric Exponential Tilting (Internal Generic)
}
\keyword{internal}
