% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmar.R
\name{nmar}
\alias{nmar}
\title{Testowa nazwa}
\usage{
nmar()
}
\arguments{
\item{x}{The entire data frame in a matrix form}

\item{col_y}{name of y (target) column}

\item{cols_y_observed}{Columns that affect y density (y ~ col_y_observed)}

\item{cols_delta}{Columns that affect observation probability (aprt from y)}

\item{prob_model_type}{Response probability model}

\item{y_dens}{Density y ~ col_y_observed}

\item{tol_value}{Tolerance in convergence step}

\item{min_iter}{Minimum iterations to nleqslv}

\item{max_iter}{Maximum iterations to nleqslv}

\item{optim_method}{Optim method to nleqslv}
}
\description{
Testowa nazwa
}
