% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_nmar_result_print.R
\name{print.nmar_result}
\alias{print.nmar_result}
\title{Print method for nmar_result}
\usage{
\method{print}{nmar_result}(x, ...)
}
\arguments{
\item{x}{nmar_result object}

\item{...}{Additional parameters}
}
\value{
`x`, invisibly.
}
\description{
Print method for nmar_result
}
\keyword{result_view}
