% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_nmar_result_print.R
\name{print.summary_nmar_result}
\alias{print.summary_nmar_result}
\title{Print method for summary.nmar_result}
\usage{
\method{print}{summary_nmar_result}(x, ...)
}
\arguments{
\item{x}{summary_nmar_result object}

\item{...}{Additional parameters}
}
\description{
Print method for summary.nmar_result
}
\keyword{result_view}
