% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_families.R
\name{probit_family}
\alias{probit_family}
\title{Probit family functions (link and derivatives)}
\usage{
probit_family()
}
\description{
Probit family functions (link and derivatives)
}
\keyword{internal}
