% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engines_el_run_engine.R
\name{run_engine.nmar_engine_el}
\alias{run_engine.nmar_engine_el}
\title{Run method for EL engine}
\usage{
\method{run_engine}{nmar_engine_el}(engine, formula, data, trace_level = 0)
}
\arguments{
\item{engine}{An object of class \code{nmar_engine_el}.}

\item{formula}{A two-sided formula passed through by \code{nmar()}.}

\item{data}{A \code{data.frame} or \code{survey.design}.}

\item{trace_level}{Integer 0-3 controlling verbosity.}
}
\value{
An object of class \code{nmar_result_el}.
}
\description{
Run method for EL engine
}
\keyword{internal}
