% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_numerics.R
\name{sanitize_nleqslv_control}
\alias{sanitize_nleqslv_control}
\title{Parse nleqslv control list for compatibility}
\usage{
sanitize_nleqslv_control(ctrl)
}
\description{
Parse nleqslv control list for compatibility
}
\keyword{internal}
