% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_nmar_result_print.R
\name{summary.nmar_result}
\alias{summary.nmar_result}
\title{Summary method for nmar_result}
\usage{
\method{summary}{nmar_result}(object, conf.level = 0.95, ...)
}
\arguments{
\item{object}{nmar_result object}

\item{conf.level}{Confidence level for intervals.}

\item{...}{Additional parameters}
}
\value{
An object of class `summary_nmar_result`.
}
\description{
Summary method for nmar_result
}
\keyword{result_view}
