% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_nmar_result_methods.R
\name{vcov.nmar_result}
\alias{vcov.nmar_result}
\title{Variance-covariance for NMAR results}
\usage{
\method{vcov}{nmar_result}(object, ...)
}
\arguments{
\item{object}{An object of class `nmar_result`.}

\item{...}{Ignored.}
}
\value{
A 1x1 numeric matrix (the variance of the primary estimate).
}
\description{
Variance-covariance for NMAR results
}
\keyword{result_param}
